# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AMS Foundation API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/demo": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Demo",
        "operationId": "demo_api_auth_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invitations/accept": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Accept",
        "operationId": "accept_api_invitations_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/academies": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Academies",
        "operationId": "academies_api_platform_academies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AcademyDto"
                  },
                  "type": "array",
                  "title": "Response Academies Api Platform Academies Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Create Academy",
        "operationId": "create_academy_api_platform_academies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademyCreatedDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/academies/{academy_id}": {
      "patch": {
        "tags": [
          "platform"
        ],
        "summary": "Set Active",
        "operationId": "set_active_api_platform_academies__academy_id__patch",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademyDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/dashboard": {
      "get": {
        "tags": [
          "academy"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_academies__academy_id__dashboard_get",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/branches": {
      "get": {
        "tags": [
          "academy"
        ],
        "summary": "Branches",
        "operationId": "branches_api_academies__academy_id__branches_get",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BranchDto"
                  },
                  "title": "Response Branches Api Academies  Academy Id  Branches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "academy"
        ],
        "summary": "Create Branch",
        "operationId": "create_branch_api_academies__academy_id__branches_post",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/branches/{branch_id}/tables": {
      "post": {
        "tags": [
          "academy"
        ],
        "summary": "Create Table",
        "operationId": "create_table_api_academies__academy_id__branches__branch_id__tables_post",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          },
          {
            "name": "branch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Branch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/members": {
      "get": {
        "tags": [
          "academy"
        ],
        "summary": "Members",
        "operationId": "members_api_academies__academy_id__members_get",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDto"
                  },
                  "title": "Response Members Api Academies  Academy Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/members/{member_id}": {
      "patch": {
        "tags": [
          "academy"
        ],
        "summary": "Update Member",
        "operationId": "update_member_api_academies__academy_id__members__member_id__patch",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/invitations": {
      "get": {
        "tags": [
          "academy"
        ],
        "summary": "Invitations",
        "operationId": "invitations_api_academies__academy_id__invitations_get",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDto"
                  },
                  "title": "Response Invitations Api Academies  Academy Id  Invitations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "academy"
        ],
        "summary": "Invite",
        "operationId": "invite_api_academies__academy_id__invitations_post",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationCreatedDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/invitations/{invitation_id}/revoke": {
      "post": {
        "tags": [
          "academy"
        ],
        "summary": "Revoke",
        "operationId": "revoke_api_academies__academy_id__invitations__invitation_id__revoke_post",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/academies/{academy_id}/activity": {
      "get": {
        "tags": [
          "academy"
        ],
        "summary": "Activity",
        "operationId": "activity_api_academies__academy_id__activity_get",
        "parameters": [
          {
            "name": "academy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Academy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditDto"
                  },
                  "title": "Response Activity Api Academies  Academy Id  Activity Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcademyCreatedDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "invitationUrl": {
            "type": "string",
            "title": "Invitationurl"
          },
          "academy": {
            "$ref": "#/components/schemas/AcademyDto"
          }
        },
        "type": "object",
        "required": [
          "id",
          "invitationUrl",
          "academy"
        ],
        "title": "AcademyCreatedDto"
      },
      "AcademyDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "active",
          "timezone",
          "createdAt"
        ],
        "title": "AcademyDto"
      },
      "AcademyInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug"
          },
          "adminEmail": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Adminemail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "slug",
          "adminEmail"
        ],
        "title": "AcademyInput"
      },
      "AcceptDto": {
        "properties": {
          "academyId": {
            "type": "string",
            "format": "uuid",
            "title": "Academyid"
          }
        },
        "type": "object",
        "required": [
          "academyId"
        ],
        "title": "AcceptDto"
      },
      "AcceptInput": {
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "AcceptInput"
      },
      "ActiveInput": {
        "properties": {
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "active"
        ],
        "title": "ActiveInput"
      },
...
      "BranchDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableDto"
            },
            "type": "array",
            "title": "Tables"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "city",
          "address",
          "tables"
        ],
        "title": "BranchDto"
      },
      "BranchInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "City"
          },
          "address": {
            "type": "string",
            "maxLength": 300,
            "minLength": 3,
            "title": "Address"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "city",
          "address"
        ],
        "title": "BranchInput"
      },
      "DashboardDto": {
        "properties": {
          "branches": {
            "type": "integer",
            "title": "Branches"
          },
          "tables": {
            "type": "integer",
            "title": "Tables"
          },
          "members": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Members"
          },
          "invitations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invitations"
          }
        },
        "type": "object",
        "required": [
          "branches",
          "tables",
          "members",
          "invitations"
        ],
        "title": "DashboardDto"
      },
      "DemoInput": {
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "userId"
        ],
        "title": "DemoInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvitationCreatedDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "invitationUrl": {
            "type": "string",
            "title": "Invitationurl"
          }
        },
        "type": "object",
        "required": [
          "id",
          "invitationUrl"
        ],
        "title": "InvitationCreatedDto"
      },
...
      "MemberDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "type": "array",
            "title": "Roles"
          },
          "allBranches": {
            "type": "boolean",
            "title": "Allbranches"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "branchIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Branchids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "name",
          "email",
          "roles",
          "allBranches",
          "active",
          "branchIds"
        ],
        "title": "MemberDto"
      },
      "MemberInput": {
        "properties": {
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 1,
            "title": "Roles"
          },
          "allBranches": {
            "type": "boolean",
            "title": "Allbranches"
          },
          "branchIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Branchids"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "roles",
          "allBranches",
          "branchIds",
          "active"
        ],
        "title": "MemberInput"
      },
...
      "TableDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TableDto"
      },
      "TableInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TableInput"
      },
      "TokenDto": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          }
        },
        "type": "object",
        "required": [
          "accessToken"
        ],
        "title": "TokenDto"
      },
...
    }
  }
}